correccion display C_interrupt

+++ Lab4/C_Interrupt/include/7_seg.h
@@ -67,6 +67,6 @@
      * 
      * @return void
      */
-    void write_decimals(uint8_t value,uint8_t*run);
+    void write_decimals(uint16_t value,uint8_t*run);
 
 #endif
